+++ lib/calendar_view.dart
@@ -24,4 +24,7 @@ export './src/theme/week_view_theme_data.dart';
 export './src/theme/multi_day_view_theme_data.dart';
 export './src/typedefs.dart';
 export './src/week_view/week_view.dart';
+export './src/month_view/month_view_style.dart';
+export './src/month_view/month_view_theme_settings.dart';
+export './src/month_view/month_view_builders.dart';
 export './src/multi_day_view/multi_day_view.dart';