+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fixed resizing the window automatically scrolls the page to the top in 
   example app. [#480](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/480)
 - Adds support for dark theme. [#263](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/263)
+- Add `MonthViewBuilders` and `MonthViewStyle` for better customisation of month view. [#501](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/pull/501)
 
 # [1.4.0 - 7 Jan 2025](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/tree/1.4.0)